backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
websockets>=12.0
diffusers>=0.31
transformers>=4.45
torch>=2.4
accelerate>=0.34
safetensors
sentencepiece
Pillow>=10.0
pydantic-settings>=2.0
